MiliApiLMS/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
greenlet==3.0.3
ImageHash==4.3.1
imageio==2.34.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pip==24.0
pluggy==1.4.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pymssql==2.2.11
pyodbc==5.1.0
pyparsing==3.1.1
pytest==8.0.2
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
scikit-image==0.22.0
scipy==1.12.0
setuptools==69.1.1
six==1.16.0
SQLAlchemy==2.0.27
tifffile==2024.2.12
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
